+++ src/whatsapp_media_organizer/__init__.py
@@ -2,7 +2,9 @@
 
 from .core import (
     OrganizeResult,
+    extract_archive,
     extract_date,
+    is_archive,
     is_valid_date,
     organize_whatsapp_media,
     unique_destination,
@@ -12,7 +14,9 @@ from .version import __version__
 __all__ = [
     "OrganizeResult",
     "__version__",
+    "extract_archive",
     "extract_date",
+    "is_archive",
     "is_valid_date",
     "organize_whatsapp_media",
     "unique_destination",